backend/requirements.txt
# PocketCare Backend Dependencies

# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# Authentication
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# Input Validation
email-validator==2.1.0

# HTTP Requests (for API calls)
requests==2.31.0

# Google Gemini AI
google-generativeai==0.3.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Production Server
gunicorn==21.2.0

# Mysql Connector 
	# Removed duplicate entry
mysql-connector-python==8.1.0

